Fix get available bands
